Move update routine to migration class
